refactor(ui): dedupe data-table state hooks and split responsive modules
- Share getDefaultDataTableRowId, useStableGetRowId, buildUseDataTableStateReturn
- Narrow useDataTableStateCore options; restore deprecated useDataTableState alias
  (core only; URL sync requires useDataTableStateWithUrl per Rules of Hooks)
- Extract DataTableBody*, responsive chrome/inner, and shared props types
- Document searchColumnId vs URL searchKey, wrapper defaults, remount behavior
